GW-80 fix multiple answers bug
Hibernate may fail if there are multiple user answers in the table
but the annotation says OneToOne. Change annotation to OneToMany,
user answers is now a collection and we always take the first
element. There should be only one element anyway under normal
circumstances. It turns out that it is somehow possible to save
multiple answers to the same question, but it is a separate bug.